chore: made hiiq table a page and added layout for stake and hiiq
