+++ src/app/dashboard/(hiiq)/hiiq-holders/page.tsx
@@ -1,0 +1,10 @@
+import HIIQHolders from '@/components/client/HIIQHolders'
+import React from 'react'
+
+export default function page() {
+  return (
+    <div>
+      <HIIQHolders />
+    </div>
+  )
+}
